requirements.txt
streamlit
st-gsheets-connection
pandas
plotly
python-dateutil
